requirements.txt
python-telegram-bot==20.4
SQLAlchemy==2.0.19
aiosqlite==0.19.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
Pillow==10.0.0
aiofiles==23.1.0
